requirements.txt
pytest==8.3.4
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
openai>=1.99.0
python-multipart==0.0.20
python-docx==1.1.2
pypdf==5.1.0
